requirements.txt
###### Requirements without Version Specifiers ######`
urllib3
PyYAML
requests

###### Requirements with Version Specifiers ######`
#certifi==2019.11.28
#chardet==3.0.4
#dbus-python==1.2.16
#distro-info===0.23ubuntu1
#idna==2.8
#PyGObject==3.36.0
#python-apt==2.0.0+ubuntu0.20.4.7
#python-dotenv==0.20.0
#requests-unixsocket==0.2.0
#six==1.14.0
#unattended-upgrades==0.1
